demo/requirements.txt
streamlit
ultralytics
torch
torchvision
opencv-python
numpy
Pillow